Converted functionality to a trait Kept the FormBuilder intact in case you need it, but now it's possible to with other extensions of the FormBuilder
